BLD: Moved SGX logic into enigma_tools_u
