+++ enigma-tools-u/Cargo.toml
@@ -16,7 +16,7 @@ etcommon-rlp = "0.2.3"
 base64 = "0.9.2"
 openssl = "0.10.10"
 rustc-hex = "1.0.0"
-# web3 utils 
+# web3 utils
 web3 = { git = 'https://github.com/tomusdrw/rust-web3.git' }
 libc = "0.2.41"
 tiny-keccak = "1.4.2"
